Move checkSerializable to Util
