+++ hazelcast-jet-core/src/test/java/com/hazelcast/jet/function/DistributedComparatorTest.java
@@ -23,7 +23,7 @@ import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
 
-import static com.hazelcast.jet.stream.impl.StreamUtil.checkSerializable;
+import static com.hazelcast.jet.impl.util.Util.checkSerializable;
 
 @Category(QuickTest.class)
 @RunWith(HazelcastParallelClassRunner.class)
